Reduce / Remove Ghosting
After a lot of trial and error, I believe I've found the resolution to the ghosting, and it's as simple as reducing the brightness level to 60. The actual brightness doesn't change it seems, but the timing of the OE pin being set HIGH is performed earlier which can work better on some panels.

Closes: #14 #12