Add a note on how to setup with Adafruit boards
Hey, amazing library!

I struggled for a while to work with my Adafruit MatrixPortal S3, I'm totally new to this so I had no idea where to find the pinouts. Adding something like this to the README would have saved me an hour :)

I'm not sure if there's a better canonical source for the pinout info than the Adafruit example source code? Happy to change!

Thanks :)