Fix double buffering #755
The existing way of delaying the timing the 'flip' to the back-buffer based on an ESP32 hardware level interrupt doesn't work properly for a variety of reasons, namely that the interrupt takes too long to process. As such, when flipDMABuffer() is called, it happens immediately. It is up to the user to ensure that they don't flip so quickly that they introduce flicker from excessive flipping!

Double-buffer example updated accordingly.